Moved some scripts to macros
